Move opm functions from build an utils to opm_operations file
