+++ iib/workers/tasks/build.py
@@ -32,11 +32,13 @@ from iib.workers.tasks.opm_operations import (
     opm_registry_rm_fbc,
     deprecate_bundles_fbc,
     generate_cache_locally,
+    opm_index_add,
+    opm_index_rm,
+    deprecate_bundles,
 )
 from iib.workers.tasks.utils import (
     add_max_ocp_version_property,
     chmod_recursively,
-    deprecate_bundles,
     get_bundles_from_deprecation_list,
     get_bundle_json,
     get_resolved_bundles,
@@ -453,143 +455,6 @@ def _get_missing_bundles(
     return filtered_bundles
 
 
-@retry(
-    before_sleep=before_sleep_log(log, logging.WARNING),
-    reraise=True,
-    retry=retry_if_exception_type(IIBError),
-    stop=stop_after_attempt(2),
-)
-def _opm_index_add(
-    base_dir: str,
-    bundles: List[str],
-    binary_image: str,
-    from_index: Optional[str] = None,
-    graph_update_mode: Optional[str] = None,
-    overwrite_from_index_token: Optional[str] = None,
-    overwrite_csv: bool = False,
-    container_tool: Optional[str] = None,
-) -> None:
-    """
-    Add the input bundles to an operator index.
-
-    This only produces the index.Dockerfile file and does not build the container image.
-
-    :param str base_dir: the base directory to generate the database and index.Dockerfile in.
-    :param list bundles: a list of strings representing the pull specifications of the bundles to
-        add to the index image being built.
-    :param str binary_image: the pull specification of the container image where the opm binary
-        gets copied from. This should point to a digest or stable tag.
-    :param str from_index: the pull specification of the container image containing the index that
-        the index image build will be based from.
-    :param str graph_update_mode: Graph update mode that defines how channel graphs are updated
-        in the index.
-    :param str overwrite_from_index_token: the token used for overwriting the input
-        ``from_index`` image. This is required to use ``overwrite_from_index``.
-        The format of the token must be in the format "user:password".
-    :param bool overwrite_csv: a boolean determining if a bundle will be replaced if the CSV
-        already exists.
-    :param str container_tool: the container tool to be used to operate on the index image
-    :raises IIBError: if the ``opm index add`` command fails.
-    """
-    # The bundles are not resolved since these are stable tags, and references
-    # to a bundle image using a digest fails when using the opm command.
-    bundle_str = ','.join(bundles) or '""'
-    cmd = [
-        'opm',
-        'index',
-        'add',
-        # This enables substitutes-for functionality for rebuilds. See
-        # https://github.com/operator-framework/enhancements/blob/master/enhancements/substitutes-for.md
-        '--enable-alpha',
-        '--generate',
-        '--bundles',
-        bundle_str,
-        '--binary-image',
-        binary_image,
-    ]
-    if container_tool:
-        cmd.append('--container-tool')
-        cmd.append(container_tool)
-
-    if graph_update_mode:
-        log.info('Using %s mode to update the channel graph in the index', graph_update_mode)
-        cmd.extend(['--mode', graph_update_mode])
-
-    log.info('Generating the database file with the following bundle(s): %s', ', '.join(bundles))
-    if from_index:
-        log.info('Using the existing database from %s', from_index)
-        # from_index is not resolved because podman does not support digest references
-        # https://github.com/containers/libpod/issues/5234 is filed for it
-        cmd.extend(['--from-index', from_index])
-
-    if overwrite_csv:
-        log.info('Using force to add bundle(s) to index')
-        cmd.extend(['--overwrite-latest'])
-
-    with set_registry_token(overwrite_from_index_token, from_index, append=True):
-        run_cmd(cmd, {'cwd': base_dir}, exc_msg='Failed to add the bundles to the index image')
-
-
-@retry(
-    before_sleep=before_sleep_log(log, logging.WARNING),
-    reraise=True,
-    retry=retry_if_exception_type(IIBError),
-    stop=stop_after_attempt(2),
-)
-def _opm_index_rm(
-    base_dir: str,
-    operators: List[str],
-    binary_image: str,
-    from_index: str,
-    overwrite_from_index_token: Optional[str] = None,
-    container_tool: Optional[str] = None,
-) -> None:
-    """
-    Remove the input operators from the operator index.
-
-    This only produces the index.Dockerfile file and does not build the container image.
-
-    :param str base_dir: the base directory to generate the database and index.Dockerfile in.
-    :param list operators: a list of strings representing the names of the operators to
-        remove from the index image.
-    :param str binary_image: the pull specification of the container image where the opm binary
-        gets copied from.
-    :param str from_index: the pull specification of the container image containing the index that
-        the index image build will be based from.
-    :param str overwrite_from_index_token: the token used for overwriting the input
-        ``from_index`` image. This is required to use ``overwrite_from_index``.
-        The format of the token must be in the format "user:password".
-    :param str container_tool: the container tool to be used to operate on the index image
-    :raises IIBError: if the ``opm index rm`` command fails.
-    """
-    cmd = [
-        'opm',
-        'index',
-        'rm',
-        '--generate',
-        '--binary-image',
-        binary_image,
-        '--from-index',
-        from_index,
-        '--operators',
-        ','.join(operators),
-    ]
-
-    if container_tool:
-        cmd.append('--container-tool')
-        cmd.append(container_tool)
-
-    log.info(
-        'Generating the database file from an existing database %s and excluding'
-        ' the following operator(s): %s',
-        from_index,
-        ', '.join(operators),
-    )
-
-    with set_registry_token(overwrite_from_index_token, from_index, append=True):
-        run_cmd(cmd, {'cwd': base_dir}, exc_msg='Failed to remove operators from the index image')
-
-
 def _overwrite_from_index(
     request_id: int,
     output_pull_spec: str,
@@ -970,7 +835,7 @@ def handle_add_request(
                 overwrite_csv=(prebuild_info['distribution_scope'] in ['dev', 'stage']),
             )
         else:
-            _opm_index_add(
+            opm_index_add(
                 base_dir=temp_dir,
                 bundles=resolved_bundles,
                 binary_image=prebuild_info['binary_image_resolved'],
@@ -1225,7 +1090,7 @@ def handle_rm_request(
             generate_cache_locally(temp_dir, fbc_dir_path, local_cache_path)
 
         else:
-            _opm_index_rm(
+            opm_index_rm(
                 base_dir=temp_dir,
                 operators=operators,
                 binary_image=prebuild_info['binary_image'],
